fix issue#62

+++ contracts/Vault.sol
@@ -106,7 +106,7 @@ contract Vault is IVault {
 
         balance += _amount;
         totalAttributions += _attributions;
-        for (uint128 i = 0; i < 2; i++) {
+        for (uint256 i = 0; i < 2; i++) {
             uint256 _allocation = (_shares[i] * _attributions) / MAGIC_SCALE_1E6;
             attributions[_beneficiaries[i]] += _allocation;
             _allocations[i] = _allocation;
